requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
torch==2.1.2
transformers==4.36.0
tokenizers==0.15.0
sentence-transformers==2.3.1
faiss-cpu>=1.9.0,<1.10.0
numpy>=1.25.0,<1.27.0
tqdm==4.66.1
typing-extensions>=4.8.0
psutil==5.9.8